03_hacker/requirements.txt
beautifulsoup4==4.12.2
PyYAML==6.0.1
redis==4.6.0
